+++ PgRoutiner/DataAccess/GetSequences.cs
@@ -31,6 +31,7 @@ public static partial class DataAccessConnectionExtensions
         {
             Schema = t.Schema,
             Name = t.Name,
+            TypeName = "SEQ",
             Type = PgType.Sequence
         });
     }